% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sri_palette.R
\name{sri_palette}
\alias{sri_palette}
\title{An SRI palette generator}
\usage{
sri_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{sricolors}, \code{sricolors2},  \code{sricool},
\code{srisunset}, \code{srioranges},  \code{sriblues}}

\item{n}{Number of colors desired.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are custom SRI color palettes
}
\examples{
sri_palette("sricolors")
sri_palette("sricool")
sri_palette("srisunset")

ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
geom_point(size = 3) +
scale_color_manual(values = sri_palette("sricolors"))


# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- sri_palette(12, name = "sricool", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}
