% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sri_palette.R
\docType{data}
\name{sri_palettes}
\alias{sri_palettes}
\title{list of custom SRI color palettes}
\format{An object of class \code{list} of length 6.}
\usage{
sri_palettes
}
\description{
Use \code{\link{sri_palettes}} to construct palettes of desired length.
}
\keyword{datasets}
