% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sri.R
\docType{data}
\name{sribasic}
\alias{sribasic}
\title{SRI basic theme and colors}
\format{An object of class \code{list} of length 3.}
\usage{
sribasic
}
\description{
Apply basic theme and colors to plots. This is a combination of theme_sri and sets the default color palette to "sricolors".
Use for simple plots only with 1-5 qualitative colors
}
\examples{
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
geom_point(size = 3) +
sribasic
}
\keyword{datasets}
