% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sri.R
\name{Theme_SRI}
\alias{Theme_SRI}
\title{SRI theme for plots}
\usage{
Theme_SRI(base_size = 11, font = "Helvetica")
}
\arguments{
\item{font}{Helvetica}
}
\description{
This is a modification of theme_minimal
}
\examples{
Theme_sri()
}
\keyword{SRI}
\keyword{theme}
