% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sri.R
\name{theme_sri}
\alias{theme_sri}
\title{SRI theme for plots}
\usage{
theme_sri(base_size = 11, font = "Helvetica")
}
\arguments{
\item{font}{Helvetica}
}
\description{
This is a modification of theme_minimal
}
\examples{
theme_sri()
}
\keyword{SRI}
\keyword{theme}
